+++ examples/load_in_mujoco.py
@@ -22,12 +22,8 @@ This example requires MuJoCo, which is installed by ``pip install mujoco``.
 """
 
 import argparse
-from importlib import import_module  # type: ignore
 
-try:
-    import mujoco
-except ImportError as e:
-    raise ImportError("MuJoCo not found, try ``pip install mujoco``") from e
+from robot_descriptions.loaders.mujoco import load_robot_description
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description=__doc__)
@@ -35,14 +31,8 @@ if __name__ == "__main__":
     args = parser.parse_args()
 
     try:
-        module = import_module(f"robot_descriptions.{args.name}")
+        model = load_robot_description(args.name)
     except ModuleNotFoundError:
-        module = import_module(f"robot_descriptions.{args.name}_description")
+        model = load_robot_description(f"{args.name}_description")
 
-    if not hasattr(module, "MJCF_PATH"):
-        raise ValueError(f"{args.name} is not an MJCF description")
-
-    model = mujoco.MjModel.from_xml_path(module.MJCF_PATH)
-    data = mujoco.MjData(model)
-
-    print(f"Robot successfully loaded with model={model} and data={data}")
+    print(f"Robot successfully loaded with model={model}")