+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Added
+
+- Loader: MuJoCo
+
 ## [0.4.0] - 2022/09/14
 
 ### Added